+++ src/main/kotlin/domain/KontorEndringsType.kt
@@ -3,6 +3,15 @@ package no.nav.domain
 enum class KontorEndringsType {
     AutomatiskRutetTilNOE,
     FlyttetAvVeileder,
-    BleSkjermer,
+    BleSkjermet,
     FikkAddressebeskyttelse,
+}
+
+fun KontorEndringsType.getKilde(): KontorKilde {
+    return when (this) {
+        KontorEndringsType.AutomatiskRutetTilNOE -> KontorKilde.ARBEIDSOPPFOLGING
+        KontorEndringsType.FlyttetAvVeileder -> KontorKilde.ARBEIDSOPPFOLGING
+        KontorEndringsType.BleSkjermet -> KontorKilde.GEOGRAFISK_TILKNYTNING
+        KontorEndringsType.FikkAddressebeskyttelse -> KontorKilde.GEOGRAFISK_TILKNYTNING
+    }
 }
