+++ src/main/kotlin/domain/KontorKilde.kt
@@ -1,0 +1,7 @@
+package no.nav.domain
+
+enum class KontorKilde {
+    ARBEIDSOPPFOLGING,
+    ARENA,
+    GEOGRAFISK_TILKNYTNING
+}
