+++ src/main/kotlin/graphql/queries/KontorQuery.kt
@@ -2,22 +2,22 @@ package no.nav.graphql.queries
 
 import com.expediagroup.graphql.server.operations.Query
 import graphql.schema.DataFetchingEnvironment
+import no.nav.db.Fnr
 import no.nav.db.table.ArbeidsOppfolgingKontorTable
 import no.nav.db.table.ArenaKontorTable
 import no.nav.db.table.GeografiskTilknytningKontorTable
+import no.nav.domain.KontorKilde
 import no.nav.graphql.schemas.KontorQueryDto
 import org.jetbrains.exposed.sql.*
 import org.jetbrains.exposed.sql.SqlExpressionBuilder.eq
 import org.jetbrains.exposed.sql.transactions.transaction
-
-enum class KontorKilde {ARBEIDSOPPFOLGING, ARENA, GEOGRAFISK_TILKNYTNING}
 
 val kontorAlias = ArbeidsOppfolgingKontorTable.kontorId.alias("kontorid")
 val kontorkildeAlias = stringLiteral(KontorKilde.ARBEIDSOPPFOLGING.name).alias("kilde") // Tilfeldig valgt verdi
 val prioritetAlias = intLiteral(0).alias("prioritet")
 
 class KontorQuery : Query {
-    fun kontorForBruker(fnrParam: String, dataFetchingEnvironment: DataFetchingEnvironment): KontorQueryDto? {
+    fun kontorForBruker(fnrParam: Fnr, _: DataFetchingEnvironment): KontorQueryDto? {
         return transaction {
 
             val arbeidsoppfolgingKontorQuery = ArbeidsOppfolgingKontorTable.select(
@@ -50,11 +50,6 @@ class KontorQuery : Query {
                 .firstOrNull()
 
             resultRow?.let { row -> KontorQueryDto(row[kontorAlias], KontorKilde.valueOf(row[kontorkildeAlias])) }
-
         }
     }
 }
-
-class FnrParam(
-    val fnr: String,
-)
